refactor:music모듈 파일구조 변경
