+++ TREE.md
@@ -1,6 +1,6 @@
 # 📁 프로젝트 파일 구조
 
-> 마지막 업데이트: 2026-01-13 (리팩토링 완료)
+> 마지막 업데이트: 2026-01-13 (앱 내부 모듈화 완료)
 
 ```
 Backend/
@@ -19,12 +19,29 @@ Backend/
 │   ├── admin.py             # Django Admin 등록
 │   ├── apps.py              # 앱 설정
 │   ├── models.py            # 데이터 모델 (Music, Artists, Albums 등)
-│   ├── serializers.py       # DRF Serializers (JSON 직렬화)
-│   ├── views.py             # API 뷰 (검색, 상세, 인증 등)
 │   ├── urls.py              # URL 라우팅
-│   ├── services.py          # 외부 API 서비스 (iTunes API)
 │   ├── tests.py             # 테스트
-│   └── migrations/          # DB 마이그레이션
+│   │
+│   ├── 📂 serializers/      # Serializers 모듈 (JSON 직렬화)
+│   │   ├── __init__.py      # 모든 Serializer export
+│   │   ├── base.py          # 기본 (Artist, Album, Tag, AiInfo)
+│   │   ├── music.py         # 음악 관련 (MusicDetail, MusicLike)
+│   │   ├── search.py        # 검색 관련 (iTunesSearchResult)
+│   │   └── auth.py          # 인증 관련 (UserRegister, UserLogin)
+│   │
+│   ├── 📂 views/            # Views 모듈 (API 엔드포인트)
+│   │   ├── __init__.py      # 모든 View export
+│   │   ├── common.py        # 공통 유틸 (MusicPagination)
+│   │   ├── auth.py          # 인증 관련 (Register, Login)
+│   │   ├── likes.py         # 좋아요 관련 (MusicLike)
+│   │   ├── search.py        # 검색 관련 (MusicSearch)
+│   │   └── music.py         # 음악 상세 관련 (MusicDetail)
+│   │
+│   ├── 📂 services/         # 외부 API 서비스
+│   │   ├── __init__.py      # 모든 Service export
+│   │   └── itunes.py        # iTunes API 통합
+│   │
+│   └── 📂 migrations/       # DB 마이그레이션
 │
 ├── 🐳 Dockerfile             # Docker 이미지 빌드 설정
 ├── 🐳 docker-compose.yml     # 멀티 컨테이너 오케스트레이션
@@ -33,8 +50,6 @@ Backend/
 ├── 📋 README.md              # 프로젝트 설명서
 ├── 📋 TREE.md                # 파일 구조 (현재 파일)
 ├── 📋 ITUNES_API_GUIDE.md    # iTunes API 통합 가이드
-│
-├── 🧪 test_itunes_api.py     # iTunes API 통합 테스트 스크립트
 │
 ├── 🔒 .env                   # 환경 변수 (Git 제외)
 ├── 🔒 .env.example           # 환경 변수 템플릿 (팀 공유용)
@@ -47,20 +62,35 @@ Backend/
 - [x] **리팩토링**: config 폴더 구조로 정리
 - [x] **Phase 2**: 인증 및 핵심 도메인 (User, Music, Playlist)
 - [x] **Phase 3-1**: iTunes API 통합 (검색 우선 구조)
+- [x] **앱 모듈화**: views/, serializers/, services/ 폴더 구조화
 - [ ] **Phase 3-2**: 외부 API (LRCLIB) 및 비동기 작업 (Celery)
 - [ ] **Phase 4**: 데이터 시각화 및 최적화 (play_log, 차트)
 - [ ] **Phase 5**: 클라우드 이관 (AWS RDS, MQ, EC2)
 
 ## 📝 주요 변경사항
 
+### 2026-01-13 - 앱 내부 모듈화
+- ✅ `music/views/` 폴더 생성 (auth, likes, search, music)
+- ✅ `music/serializers/` 폴더 생성 (base, music, search, auth)
+- ✅ `music/services/` 폴더 생성 (itunes)
+- ✅ `__init__.py`에서 모든 클래스 export (기존 import 호환)
+- ✅ 기능별 파일 분리로 협업 충돌 감소
+
 ### 2026-01-13 - iTunes API 통합
-- ✅ iTunes Search API 서비스 구현 (`music/services.py`)
-- ✅ 고급 검색 문법 지원 (검색어 + 태그: `아이유#christmas`)
+- ✅ iTunes Search API 서비스 구현
+- ✅ 고급 검색 문법 지원 (검색어 + 태그: `아이유 # christmas`)
+- ✅ '# ' (해시+공백) 패턴만 태그로 인식 (C#, I'm #1 등 안전 처리)
 - ✅ 자동 DB 저장 (클릭 시 iTunes → DB 자동 저장)
 - ✅ AI 필터링 (`exclude_ai` 파라미터)
-- ✅ 새로운 API 엔드포인트:
-  - `GET /api/v1/search?q={검색어}` - iTunes 기반 검색
-  - `GET /api/v1/tracks/{itunes_id}` - 상세 조회 (자동 저장)
+
+### API 엔드포인트
+- `GET /api/v1/search?q={검색어}` - iTunes 기반 검색
+- `GET /api/v1/tracks/{itunes_id}` - 상세 조회 (자동 저장)
+- `POST /api/v1/tracks/{music_id}/likes` - 좋아요 등록
+- `DELETE /api/v1/tracks/{music_id}/likes` - 좋아요 취소
+- `POST /api/v1/auth/users/` - 회원가입
+- `POST /api/v1/auth/tokens/` - 로그인
+- `POST /api/v1/auth/refresh/` - 토큰 갱신
 
 ### 2026-01-13 - 리팩토링 완료
 - ✅ `config/` 폴더 생성 및 설정 파일 이동
